+++ Sources/Enums/PresentationAnimationType.swift
@@ -32,8 +32,9 @@ public enum PresentationAnimationType: IBEnum {
     guard let string = string else {
       return nil
     }
-
-    let (name, params) = MaskType.extractNameAndParams(from: string)
+    guard let (name, params) = string.extractNameAndParams() else {
+      return nil
+    }
     switch name {
     case "crossdissolve":
       self = .crossDissolve
@@ -53,9 +54,15 @@ public enum PresentationAnimationType: IBEnum {
 }
 
 extension PresentationAnimationType: Hashable {
+  #if swift(>=4.2)
+  public func hash(into hasher: inout Hasher) {
+    hasher.combine(stringValue)
+  }
+  #else
   public var hashValue: Int {
     return stringValue.hashValue
   }
+  #endif
 
   public static func == (lhs: PresentationAnimationType, rhs: PresentationAnimationType) -> Bool {
     return lhs.stringValue == rhs.stringValue